+++ src/main/gen/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/RobotVisitor.java
@@ -61,6 +61,10 @@ public class RobotVisitor extends PsiElementVisitor {
     // visitFoldable(o);
   }
 
+  public void visitExceptionHandlingStructure(@NotNull RobotExceptionHandlingStructure o) {
+    visitExecutableStatement(o);
+  }
+
   public void visitExecutableStatement(@NotNull RobotExecutableStatement o) {
     visitElement(o);
   }